requirements.txt
aiohttp==3.8.5
aiosignal==1.3.1
annotated-types==0.5.0
anyio==3.7.1
astor==0.8.1
async-timeout==4.0.3
asynctest==0.13.0
attrdict==2.0.1
attrs==23.1.0
Babel==2.12.1
backcall==0.2.0
bce-python-sdk==0.8.90
beautifulsoup4==4.12.2
cachetools==5.3.1
charset-normalizer==3.3.0
click==8.1.7
colorama==0.4.6
colorlog==6.7.0
cssselect==1.2.0
cssutils==2.7.1
cycler==0.11.0
Cython==3.0.2
datasets==2.13.2
debugpy==1.7.0
decorator==5.1.1
dill==0.3.4
entrypoints==0.4
et-xmlfile==1.1.0
exceptiongroup==1.1.3
fastapi==0.103.2
filelock==3.12.2
fire==0.5.0
Flask==2.2.5
flask-babel==3.1.0
fonttools==4.38.0
frozenlist==1.3.3
fsspec==2023.1.0
future==0.18.3
h11==0.14.0
httpcore==0.17.3
httpx==0.24.1
huggingface-hub==0.16.4
idna==3.4
imageio==2.31.2
imgaug==0.4.0
importlib-metadata==6.7.0
#ipykernel==6.16.2
ipython==7.34.0
itsdangerous==2.1.2
jedi==0.19.0
jieba==0.42.1
Jinja2==3.1.2
joblib==1.3.2
jupyter_client==7.4.9
jupyter_core==4.12.0
kiwisolver==1.4.5
lmdb==1.4.1
lxml==4.9.3
markdown-it-py==2.2.0
MarkupSafe==2.1.3
matplotlib==3.5.3
matplotlib-inline==0.1.6
mdurl==0.1.2
multidict==6.0.4
multiprocess==0.70.12.2
nest-asyncio==1.5.8
networkx==2.6.3
nltk==3.8.1
numpy==1.21.6
onnx==1.14.1
#opencv-contrib-python==4.6.0.66
opencv-python==4.6.0.66
openpyxl==3.1.2
opt-einsum==3.3.0
packaging==23.2
paddle-bfloat==0.1.7
paddle2onnx==1.0.6
paddlefsl==1.1.0
paddlenlp==2.6.1
paddleocr==2.7.0.2
paddlepaddle-gpu==2.5.1
pandas==1.3.5
parso==0.8.3
pdf2docx==0.5.6
pickleshare==0.7.5
Pillow==9.5.0
premailer==3.10.0
prompt-toolkit==3.0.39
protobuf==3.20.2
psutil==5.9.5
pyarrow==12.0.1
pyclipper==1.3.0.post5
pycryptodome==3.19.0
pydantic==2.4.2
pydantic_core==2.10.1
Pygments==2.16.1
PyMuPDF==1.20.2
pyparsing==3.1.1
python-dateutil==2.8.2
python-docx==0.8.11
pytz==2023.3.post1
PyWavelets==1.3.0
#pywin32==306
PyYAML==6.0.1
pyzmq==25.1.1
rapidfuzz==3.3.1
rarfile==4.1
regex==2023.8.8
requests==2.31.0
rich==13.6.0
safetensors==0.3.3
scikit-image==0.19.3
scikit-learn==1.0.2
scipy==1.7.3
sentencepiece==0.1.99
seqeval==1.2.2
shapely==2.0.1
six==1.16.0
sniffio==1.3.0
soupsieve==2.4.1
starlette==0.27.0
termcolor==2.3.0
threadpoolctl==3.1.0
tifffile==2021.11.2
tokenizers==0.13.3
torch==1.8.1
tornado==6.2
tqdm==4.66.1
traitlets==5.9.0
transformers==4.30.2
typer==0.9.0
typing_extensions==4.7.1
unicodedata2==15.1.0
Unidecode==1.3.7
urllib3==2.0.5
uvicorn==0.22.0
visualdl==2.5.3
wcwidth==0.2.8
Werkzeug==2.2.3
wincertstore==0.2
xxhash==3.3.0
yarl==1.9.2
zipp==3.15.0
